Added TypeVar for specifying input data types (int, float, str, bool)
Corrected code after code review. Added TypeVar for specifying input data types (int, float, str, bool). Small corrections by texting/comments
